funcionalidad de eliminar turno agregada
