+++ FRONTEND/Dashboard/index.html
@@ -376,7 +376,7 @@
         </div>
     </div>
 
-    <script src="script.js" type="module"></script>
+    <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
